Widget tests for ProductListTest and unit for repositories
Error: ══╡ EXCEPTION CAUGHT BY FLUTTER TEST FRAMEWORK ╞════════════════════════════════════════════════════
The following StateError was thrown running a test:
Bad state: No method stub was called from within `when()`. Was a real method called, or perhaps an
extension method?

When the exception was thrown, this was the stack:
#0      When._completeWhen (package:mocktail/src/mocktail.dart:262:7)
#1      When.thenAnswer (package:mocktail/src/mocktail.dart:257:12)
#2      whenListen (package:bloc_test/src/when_listen.dart:55:27)
#3      main.<anonymous closure> (file:///Users/kklinik/quick_sell/test/product_list_page_test.dart:24:5)
#4      testWidgets.<anonymous closure>.<anonymous closure> (package:flutter_test/src/widget_tester.dart:168:29)
<asynchronous suspension>
#5      TestWidgetsFlutterBinding._runTestBody (package:flutter_test/src/binding.dart:1013:5)
<asynchronous suspension>
<asynchronous suspension>
(elided one frame from package:stack_trace)

The test description was:
  ProductsListPage shows loading indicator